+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 10.18.0
+
+* `[knf]` Added `no` as a valid boolean value for `GetB`
+* `[knf]` Added new validators for property type validation
+* `[knf]` Code refactoring
+
 ### 10.17.0
 
 * `[cache]` Added package which provides simple in-memory key:value store